backend/requirements.txt
django>=5.1,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.3
drf-spectacular>=0.27
django-cors-headers>=4.4
django-ratelimit>=4.1
mysqlclient>=2.2
celery>=5.4
redis>=5.0
channels>=4.1
channels-redis>=4.2
gunicorn>=22.0
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
joblib>=1.4
python-dotenv>=1.0
Pillow>=10.4
reportlab>=4.2
